+++ src/main/java/com/DOH/DOH/controller/list/ContestListController.java
@@ -27,9 +27,6 @@ public class ContestListController {
     @Autowired
     private ContestListService contestListService;
 
-    @Autowired
-    private S3Service s3Service;  // S3 업로드 서비스를 추가로 주입
-
     @GetMapping("/list")
     public String contestList(@RequestParam(defaultValue = "1") int page, @RequestParam(name = "orderType", required = false, defaultValue = "current")String orderType, Model model){
 
@@ -43,45 +40,4 @@ public class ContestListController {
         return "list/contestList";
     }
 
-    @GetMapping("/application/terms")
-    public String applictionTerms(HttpSession session){
-        String email= (String) session.getAttribute("userEmail");
-        log.info("email 체크!! -> "+email);
-        return "list/applicationTerms";
-    }
-
-    @GetMapping("/application/write")
-    public String applictionWrite(String userEmail, Model model){
-        model.addAttribute("userEmail", userEmail);
-        return "list/applicationWrite";
-    }
-
-    @PostMapping("/application/upload")
-    public String handleUpload(@RequestParam("image") MultipartFile[] files,
-                               @RequestParam HashMap<String, String> param,
-                               Model model) {
-        try {
-            // 1. 파일 업로드 처리 (S3)
-            for (MultipartFile file : files) {
-                if (!file.isEmpty()) {
-                    String fileUrl = s3Service.uploadFile(file);
-                    log.info("File uploaded to S3: " + fileUrl);
-                }
-            }
-
-            // 2. 제목 등 입력값 처리 (DB)
-            ApplyDTO contest = new ApplyDTO();
-            contest.setApplyTitle(param.get("applyTitle"));
-            contest.setApplyContent(param.get("applyContent"));
-            log.info("applyTitle!!"+contest.getApplyTitle());
-            log.info("applyContent!!"+contest.getApplyContent());
-            contestListService.saveContestApply(contest);
-
-            return "redirect:/contest/list";  // 업로드 후 컨테스트 목록 페이지로 이동
-        } catch (Exception e) {
-            log.error("업로드 중 오류 발생", e);
-            model.addAttribute("error", "업로드 중 문제가 발생했습니다.");
-            return "list/applicationWrite";  // 오류 시 다시 작성 페이지로 이동
-        }
-    }
 }
