--- src/main/java/com/DOH/DOH/controller/list/FileUploadController.java
@@ -1,31 +1,0 @@
-package com.DOH.DOH.controller.list;
-import com.DOH.DOH.service.list.S3Service;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-
-@RestController
-@RequestMapping("/contest")
-public class FileUploadController {
-
-    private final S3Service s3Service;
-
-    public FileUploadController(S3Service s3Service) {
-        this.s3Service = s3Service;
-    }
-
-    @PostMapping("/upload")
-    public ResponseEntity<String> uploadFile(@RequestParam("file") MultipartFile file) {
-        try {
-            String fileUrl = s3Service.uploadFile(file);
-            return ResponseEntity.ok("File uploaded successfully: " + fileUrl);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("File upload failed.");
-        }
-    }
-
-}
