+++ src/main/resources/templates/list/applicationWrite.html
@@ -67,7 +67,7 @@
                                 <input type="file" name="" id="image" multiple>
                             </div>
                             <div class="InputArea">
-                                <div class="content text" onclick="addText()">
+                                <div class="content text" th:onclick="addText()">
                                     <span class="Image">
                                         <i class="fa-regular fa-pen-to-square"></i>
                                     </span>
